Refactor: break up frame processing among multiple file
