+++ src/transmission/processing/satellites.py
@@ -1,0 +1,12 @@
+"""Satellite related constants"""
+
+TIME_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
+
+SATELLITES = {
+    "delfi_pq": '51074',
+    "delfi_next": '39428',
+    "delfi_c3": '32789',
+    "da_vinci": None  #update id
+}
+
+
