+++ src/transmission/processing/scheduler.py
@@ -1,8 +1,8 @@
 """Scheduler for planning telemetry scrapes and frame processing"""
 from apscheduler.schedulers.background import BackgroundScheduler
 from delfipq.process_raw_telemetry import process_frames_delfi_pq
-from transmission.telemetry_scraper import scrape
-from transmission.save_data import process_uplink_and_downlink
+from transmission.processing.telemetry_scraper import scrape
+from transmission.processing.save_raw_data import process_uplink_and_downlink
 
 
 scheduler = BackgroundScheduler()
